src/requirements.txt
keras
python-magic
flask
bs4
opencv-python
tensorboard
icrawler
selenium